Adds a SNYK VERSION at build time
Instead of pulling off of /latest whenever the build action is fired, this update has the build.rb grab the tag version from latest and then set it as a static variable for the rest of the build process.

Side effect of this is whenever the build.yaml is run, it will only build with the latest base images but not change the snyk version until a new generate action is run.

This should probably be coupled with a either a watch action from here that checks for a new release against snyk cli, or dispatch action from the snyk repo to trigger a new build here.