fix: use bash instead of /bin/sh
The entrypoint script is using bash-specific syntax which fails to
execute with:

```
/usr/local/bin/docker-entrypoint.sh: 87: /usr/local/bin/docker-entrypoint.sh: Bad substitution
```

This happens on some base images, not all as `/bin/sh` is different.
They can be using busybox (alpine, works), dash (node, fails), and
others.

So the options are to use sh-only syntax and be very careful or switch
to Bash. As Bash is ubiquitous and provides more features, I've switched
the script over to use it for consistent behaviour across all images.